+++ src/net/resolver.rs
@@ -37,11 +37,11 @@ impl Iterator for SocketAddrIterator {
         let next = unsafe {
             host::api::networking::resolve_next(
                 self.id,
-                &mut addr_type as *mut u32,
-                addr.as_mut_ptr(),
-                &mut port as *mut u16,
-                &mut flowinfo as *mut u32,
-                &mut scope_id as *mut u32,
+                &mut addr_type as *mut u32 as u32,
+                addr.as_mut_ptr() as u32,
+                &mut port as *mut u16 as u16,
+                &mut flowinfo as *mut u32 as u32,
+                &mut scope_id as *mut u32 as u32,
             )
         };
 
@@ -93,10 +93,10 @@ fn resolve_timeout_(
     };
     let result = unsafe {
         host::api::networking::resolve(
-            name.as_ptr(),
-            name.len(),
+            name.as_ptr() as u32,
+            name.len() as u32,
             timeout_ms,
-            &mut dns_iter_or_error_id as *mut u64,
+            &mut dns_iter_or_error_id as *mut u64 as u64,
         )
     };
     if result != 0 {
