+++ src/net/udp.rs
@@ -37,11 +37,11 @@ impl UdpSocket {
                     unsafe {
                         host::api::networking::udp_bind(
                             4,
-                            ip.as_ptr(),
+                            ip.as_ptr() as u32,
                             port,
                             0,
                             0,
-                            &mut id as *mut u64,
+                            &mut id as *mut u64 as u64,
                         )
                     }
                 }
@@ -53,11 +53,11 @@ impl UdpSocket {
                     unsafe {
                         host::api::networking::tcp_bind(
                             6,
-                            ip.as_ptr(),
+                            ip.as_ptr() as u32,
                             port,
                             flow_info,
                             scope_id,
-                            &mut id as *mut u64,
+                            &mut id as *mut u64 as u64,
                         )
                     }
                 }
@@ -75,7 +75,10 @@ impl UdpSocket {
     pub fn local_addr(&self) -> Result<SocketAddr> {
         let mut dns_iter_or_error_id = 0;
         let result = unsafe {
-            host::api::networking::udp_local_addr(self.id, &mut dns_iter_or_error_id as *mut u64)
+            host::api::networking::udp_local_addr(
+                self.id,
+                &mut dns_iter_or_error_id as *mut u64 as u64,
+            )
         };
         if result == 0 {
             let mut dns_iter = SocketAddrIterator::from(dns_iter_or_error_id);
